chore(contentful): rename Model to Content
