+++ packages/botonic-plugin-contentful/src/cms/index.ts
@@ -1,5 +1,5 @@
 export * from './callback';
 export * from './cms';
-export * from './model';
+export * from './contents';
 export * from './cms_error';
 export * from './cms_dummy';
